+++ src/black/mode.py
@@ -193,6 +193,7 @@ class Preview(Enum):
     hug_parens_with_braces_and_square_brackets = auto()
     allow_empty_first_line_before_new_block_or_comment = auto()
     single_line_format_skip_with_multiple_comments = auto()
+    long_case_block_line_splitting = auto()
 
 
 class Deprecated(UserWarning):